#10074: Move deg2rad_bw to TTNN
