+++ docs/source/ttnn/ttnn/ttnn/deg2rad_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.deg2rad_bw:
+
+ttnn.deg2rad_bw
+###############
+
+.. autofunction:: ttnn.deg2rad_bw
