+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -147,11 +147,6 @@ std::vector<Tensor> angle_bw(
     bool is_complextensor = true,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> deg2rad_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-
 std::vector<Tensor> prod_bw(
     const Tensor& grad,
     const Tensor& input,
